refactor app upload APIs
Closes #296
