+++ platform/api/server/apply/client.go
@@ -27,20 +27,8 @@ func NewClient(instance string, logger log.Logger, token string, opts ...httptra
 		).Endpoint()
 	}
 
-	var uploadAppEndpoint endpoint.Endpoint
-	{
-		uploadAppEndpoint = httptransport.NewClient(
-			"POST",
-			copyURL(u, "/v1/apps"),
-			encodeRequestWithToken(token, EncodeUploadAppRequest),
-			DecodeUploadAppResponse,
-			opts...,
-		).Endpoint()
-	}
-
 	return Endpoints{
 		DefineDEPProfileEndpoint: defineDEPProfileEndpoint,
-		AppUploadEndpoint:        uploadAppEndpoint,
 	}, nil
 }
