+++ platform/api/server/apply/endpoint.go
@@ -2,7 +2,6 @@ package apply
 
 import (
 	"context"
-	"io"
 
 	"github.com/go-kit/kit/endpoint"
 	"github.com/micromdm/dep"
@@ -10,21 +9,6 @@ import (
 
 type Endpoints struct {
 	DefineDEPProfileEndpoint endpoint.Endpoint
-	AppUploadEndpoint        endpoint.Endpoint
-}
-
-func (e Endpoints) UploadApp(ctx context.Context, manifestName string, manifest io.Reader, pkgName string, pkg io.Reader) error {
-	request := appUploadRequest{
-		ManifestName: manifestName,
-		ManifestFile: manifest,
-		PKGFilename:  pkgName,
-		PKGFile:      pkg,
-	}
-	resp, err := e.AppUploadEndpoint(ctx, request)
-	if err != nil {
-		return err
-	}
-	return resp.(appUploadResponse).Err
 }
 
 func (e Endpoints) DefineDEPProfile(ctx context.Context, p *dep.Profile) (*dep.ProfileResponse, error) {
@@ -47,30 +31,6 @@ func MakeDefineDEPProfile(svc Service) endpoint.Endpoint {
 		}, nil
 	}
 }
-
-func MakeUploadAppEndpiont(svc Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (response interface{}, err error) {
-		req := request.(appUploadRequest)
-		err = svc.UploadApp(ctx, req.ManifestName, req.ManifestFile, req.PKGFilename, req.PKGFile)
-		return &appUploadResponse{
-			Err: err,
-		}, nil
-	}
-}
-
-type appUploadRequest struct {
-	ManifestName string
-	ManifestFile io.Reader
-
-	PKGFilename string
-	PKGFile     io.Reader
-}
-
-type appUploadResponse struct {
-	Err error `json:"err,omitempty"`
-}
-
-func (r appUploadResponse) error() error { return r.Err }
 
 type depProfileRequest struct{ *dep.Profile }
 type depProfileResponse struct {
