+++ platform/api/server/apply/transport_http.go
@@ -4,9 +4,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"io"
 	"io/ioutil"
-	"mime/multipart"
 	"net/http"
 
 	httptransport "github.com/go-kit/kit/transport/http"
@@ -26,12 +24,6 @@ func MakeHTTPHandlers(ctx context.Context, endpoints Endpoints, opts ...httptran
 			encodeResponse,
 			opts...,
 		),
-		AppUploadHandler: httptransport.NewServer(
-			endpoints.AppUploadEndpoint,
-			decodeAppUploadRequest,
-			encodeResponse,
-			opts...,
-		),
 	}
 	return h
 }
@@ -42,64 +34,6 @@ func decodeDEPProfileRequest(ctx context.Context, r *http.Request) (interface{},
 		return nil, err
 	}
 	return req, nil
-}
-
-func decodeAppUploadRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	defer r.Body.Close()
-	appManifestFilename := r.FormValue("app_manifest_filename")
-	manifestFile, _, err := r.FormFile("app_manifest_filedata")
-	if err != nil && err != http.ErrMissingFile {
-		return nil, errors.Wrap(err, "manifest file")
-	}
-	pkgFilename := r.FormValue("pkg_name")
-	pkgFile, _, err := r.FormFile("pkg_filedata")
-	if err != nil && err != http.ErrMissingFile {
-		return nil, err
-	}
-
-	return appUploadRequest{
-		ManifestName: appManifestFilename,
-		ManifestFile: manifestFile,
-		PKGFilename:  pkgFilename,
-		PKGFile:      pkgFile,
-	}, nil
-}
-
-func EncodeUploadAppRequest(_ context.Context, r *http.Request, request interface{}) error {
-	req := request.(appUploadRequest)
-	body := new(bytes.Buffer)
-	writer := multipart.NewWriter(body)
-
-	if req.ManifestName != "" {
-		partManifest, err := writer.CreateFormFile("app_manifest_filedata", req.ManifestName)
-		if err != nil {
-			return err
-		}
-		_, err = io.Copy(partManifest, req.ManifestFile)
-		if err != nil {
-			return errors.Wrap(err, "copying appmanifest file to multipart writer")
-		}
-		writer.WriteField("app_manifest_filename", req.ManifestName)
-	}
-
-	if req.PKGFilename != "" {
-		partPkg, err := writer.CreateFormFile("pkg_filedata", req.PKGFilename)
-		if err != nil {
-			return err
-		}
-		_, err = io.Copy(partPkg, req.PKGFile)
-		if err != nil {
-			return errors.Wrap(err, "copying pkg file to multipart writer")
-		}
-		writer.WriteField("pkg_name", req.PKGFilename)
-	}
-	if err := writer.Close(); err != nil {
-		return errors.Wrap(err, "closing multipart writer")
-	}
-
-	r.Header.Set("Content-Type", writer.FormDataContentType())
-	r.Body = ioutil.NopCloser(body)
-	return nil
 }
 
 type errorWrapper struct {
@@ -152,15 +86,6 @@ func DecodeDEPProfileResponse(_ context.Context, r *http.Response) (interface{},
 		return nil, errorDecoder(r)
 	}
 	var resp depProfileResponse
-	err := json.NewDecoder(r.Body).Decode(&resp)
-	return resp, err
-}
-
-func DecodeUploadAppResponse(_ context.Context, r *http.Response) (interface{}, error) {
-	if r.StatusCode != http.StatusOK {
-		return nil, errorDecoder(r)
-	}
-	var resp appUploadResponse
 	err := json.NewDecoder(r.Body).Decode(&resp)
 	return resp, err
 }
