+++ platform/api/server/list/client.go
@@ -60,23 +60,11 @@ func NewClient(instance string, logger log.Logger, token string, opts ...httptra
 		).Endpoint()
 	}
 
-	var listAppsEndpoint endpoint.Endpoint
-	{
-		listAppsEndpoint = httptransport.NewClient(
-			"GET",
-			copyURL(u, "/v1/apps"),
-			encodeRequestWithToken(token, EncodeHTTPGenericRequest),
-			DecodeListAppsResponse,
-			opts...,
-		).Endpoint()
-	}
-
 	return Endpoints{
 		ListDevicesEndpoint:       listDevicesEndpoint,
 		GetDEPAccountInfoEndpoint: getDEPAccountInfoEndpoint,
 		GetDEPDeviceEndpoint:      getDEPDeviceDetailsEndpoint,
 		GetDEPProfileEndpoint:     getDEPProfilesEndpoint,
-		ListAppsEndpont:           listAppsEndpoint,
 	}, nil
 }
