+++ platform/api/server/list/endpoint.go
@@ -13,7 +13,6 @@ type Endpoints struct {
 	GetDEPAccountInfoEndpoint endpoint.Endpoint
 	GetDEPDeviceEndpoint      endpoint.Endpoint
 	GetDEPProfileEndpoint     endpoint.Endpoint
-	ListAppsEndpont           endpoint.Endpoint
 }
 
 func (e Endpoints) ListDevices(ctx context.Context, opts ListDevicesOption) ([]DeviceDTO, error) {
@@ -23,15 +22,6 @@ func (e Endpoints) ListDevices(ctx context.Context, opts ListDevicesOption) ([]D
 		return nil, err
 	}
 	return response.(devicesResponse).Devices, response.(devicesResponse).Err
-}
-
-func (e Endpoints) ListApplications(ctx context.Context, opts ListAppsOption) ([]AppDTO, error) {
-	request := appListRequest{opts}
-	response, err := e.ListAppsEndpont(ctx, request.Opts)
-	if err != nil {
-		return nil, err
-	}
-	return response.(appListResponse).Apps, response.(appListResponse).Err
 }
 
 func (e Endpoints) GetDEPAccountInfo(ctx context.Context) (*dep.Account, error) {
@@ -59,17 +49,6 @@ func MakeListDevicesEndpoint(svc Service) endpoint.Endpoint {
 		return devicesResponse{
 			Devices: dto,
 			Err:     err,
-		}, nil
-	}
-}
-
-func MakeListAppsEndpoint(svc Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (response interface{}, err error) {
-		req := request.(appListRequest)
-		apps, err := svc.ListApplications(ctx, req.Opts)
-		return appListResponse{
-			Apps: apps,
-			Err:  err,
 		}, nil
 	}
 }
@@ -147,19 +126,3 @@ type depProfileResponse struct {
 }
 
 func (r depProfileResponse) error() error { return r.Err }
-
-type appListRequest struct {
-	Opts ListAppsOption
-}
-
-type AppDTO struct {
-	Name    string `json:"name"`
-	Payload []byte `json:"payload,omitempty"`
-}
-
-type appListResponse struct {
-	Apps []AppDTO `json:"apps,omitempty"`
-	Err  error    `json:"err,omitempty"`
-}
-
-func (r appListResponse) error() error { return r.Err }
