+++ platform/api/server/list/transport_http.go
@@ -16,7 +16,6 @@ type HTTPHandlers struct {
 	GetDEPAccountInfoHandler   http.Handler
 	GetDEPProfileHandler       http.Handler
 	GetDEPDeviceDetailsHandler http.Handler
-	ListAppsHandler            http.Handler
 }
 
 func MakeHTTPHandlers(ctx context.Context, endpoints Endpoints, opts ...httptransport.ServerOption) HTTPHandlers {
@@ -45,12 +44,6 @@ func MakeHTTPHandlers(ctx context.Context, endpoints Endpoints, opts ...httptran
 			encodeResponse,
 			opts...,
 		),
-		ListAppsHandler: httptransport.NewServer(
-			endpoints.ListAppsEndpont,
-			decodeListAppsRequest,
-			encodeResponse,
-			opts...,
-		),
 	}
 	return h
 }
@@ -76,14 +69,6 @@ func decodeDepDeviceDetailsRequest(ctx context.Context, r *http.Request) (interf
 
 func decodeDEPProfileRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	var request depProfileRequest
-	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
-		return nil, err
-	}
-	return request, nil
-}
-
-func decodeListAppsRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var request appListRequest
 	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
 		return nil, err
 	}
@@ -167,15 +152,6 @@ func DecodeDEPProfileResponse(_ context.Context, r *http.Response) (interface{},
 		return nil, errorDecoder(r)
 	}
 	var resp depProfileResponse
-	err := json.NewDecoder(r.Body).Decode(&resp)
-	return resp, err
-}
-
-func DecodeListAppsResponse(_ context.Context, r *http.Response) (interface{}, error) {
-	if r.StatusCode != http.StatusOK {
-		return nil, errorDecoder(r)
-	}
-	var resp appListResponse
 	err := json.NewDecoder(r.Body).Decode(&resp)
 	return resp, err
 }
